refactor: package cmd move test helper functions into separate test_utils.go
